+++ CHANGELOG.md
@@ -1,20 +1,24 @@
 # CHANGELOG
 
-## 0.5.0.dev
+## 0.5.0
 
 ### Sentence corpus creation
 
 - ALTO input refinements:
     - ALTO-based sentence corpora now include `title`, `author`, and `page_number` based on block tags
     - Footnotes now follow all main body content, with correct title, author, and page number metadata
+    - Zip file name is used as `file` field (for quote consolidation); individual ALTO filenames are
+        included as `page_file`
 
 ### Quotation detection
 
 - Add consolidation logic to group quotes that are sequential in both original and reuse texts
     - Quote consolidation is enabled by default but configurable in both app and command line script
+    - Quotes are only consolidated for sequential sentences from a single source file
 - Now supports multiple original sentence corpora as input
+    - `remarx-find-quotes` script accepts one or more CSV files or directories of CSVs for original corpora; uses default original corpus location when no original path is specified
 
-### Scripts
+### Non-package scripts
 
 - Add `xquery` directory with instructions and XQuery files for investigating
     & reporting on continuing paragraph and footnote content in MEGA TEI
@@ -43,7 +47,7 @@
 
 ### Quotation detection
 
-- Add the `find_quotes.py` CLI script to run the quote finder, with a `--benchmark` flag to collect performance metrics
+- Add `remarx-find-quotes` script to run the quote finder, with a `--benchmark` option to collect performance metrics
 
 ### Application